Do not discard a perfectly good line at the end of the data burst
While parsing a data burst, the inner loop consumes a line at a time and tests
if it matches the sensor data format. If it does not, the data burst is over,
so the new particle is emitted, and the function jumps back to the top o the
outer loop to start looking for the next data burst.

However, that line has now been consumed and cannot be read again by the outer
loop. If the discarded line is the header of the next data burst, parsing will
fail.

This commit fixes that by passing the line back to the outer loop.