Initial commit
Initial archive of the Dataset Parser Development Overview. This was originally a document on the OOI Google Drive. It has been updated, but still has some information which is out of date, most notably, we are using the Ingest Manager for data file ingest (although the ingest tool still works and may make it easier to debug any issues with the parser).